% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McMurrayHorstSamuelson2012-package.r
\docType{package}
\name{McMurrayHorstSamuelson2012-package}
\alias{McMurrayHorstSamuelson2012}
\alias{McMurrayHorstSamuelson2012-package}
\title{McMurrayHorstSamuelson2012: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
